Add a navigation bar for researcher's side
