--- experimental-hub-frontend/src/components/organisms/.gitignore
@@ -1,1 +1,0 @@
-!.gitignore
