+++ experimental-hub-frontend/src/components/organisms/NavigationBar/NavigationBar.css
@@ -1,0 +1,25 @@
+.headerContainer {
+    float: left;
+    background-color: #f3f3f3;
+    width: 100%;
+    display: flex;
+    height: 4.4em;
+    align-items: center;
+    position: fixed;
+    border: 1px solid #e7e7e7;
+}
+
+.pageNavigation {
+    color: #666;
+    padding: 1.4em;
+    font-size: 1.1em;
+    text-decoration: none;
+}
+
+.pageNavigation:hover {
+    background-color: #ddd;
+}
+
+.active {
+    background-color: #ddd;
+}
