--- experimental-hub-frontend/src/components/templates/.gitignore
@@ -1,1 +1,0 @@
-!.gitignore
