+++ experimental-hub-frontend/src/pages/ExperimentRoom/ExperimentRoom.js
@@ -1,4 +1,3 @@
-import { Link } from "react-router-dom";
 import "./ExperimentRoom.css";
 
 function ExperimentRoom() {
