+++ experimental-hub-frontend/src/pages/MainPage/MainPage.js
@@ -1,0 +1,11 @@
+import NavigationBar from "../../components/organisms/NavigationBar/NavigationBar";
+import "./MainPage.css";
+function MainPage() {
+  return (
+    <div>
+      <NavigationBar />
+    </div>
+  );
+}
+
+export default MainPage;
