+++ experimental-hub-frontend/src/pages/PostProcessing/PostProcessing.js
@@ -1,0 +1,12 @@
+import NavigationBar from "../../components/organisms/NavigationBar/NavigationBar";
+import "./PostProcessing.css";
+
+function PostProcessing() {
+  return (
+    <div>
+      <NavigationBar />
+    </div>
+  );
+}
+
+export default PostProcessing;
