+++ experimental-hub-frontend/src/pages/SessionOverview/SessionOverview.js
@@ -1,0 +1,12 @@
+import NavigationBar from "../../components/organisms/NavigationBar/NavigationBar";
+import "./SessionOverview.css";
+
+function SessionOverview() {
+  return (
+    <div>
+      <NavigationBar />
+    </div>
+  );
+}
+
+export default SessionOverview;
